requirements.txt
django
djangorestframework
django-cors-headers
gunicorn
whitenoise
drf_yasg
djangorestframework-simplejwt
django-allauth
requests